Add action server and clients for face detection, object and navigation.
This is a snapshot of the working demo as on the night of 11/3/2015.
The action clients are wrapped around classes that are are instantiated
in the executive layer. The reason behind doing this is to abstract
out the specifics of the actionlib implementation.

The Executive is currently a simple FSM that goes to a spot and looks
for a face. Upon finding a face it requests an object and waits for it
until the object detector says its the right one. It then thanks the
human and returns home.